[SCHEMA] Simplify transaction log construction.
This commit implements a leaner, meaner transaction log. Instead of storing an
ORM-derived delta, only store a snapshot of the API representation of the
versioned object. This reduces the volume of data we store, makes constructing
the transaction faster, and removes some complexity left over from when we were
trying to infer syncback actions from the transaction log. It's less generic,
but better suited to our actual application.

Includes some simplifications to the Lens model -- just removing parts that
would be broken by this change, and were part of webhooks but are unlikely to
be repurposed into ping.